+++ axum/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-- **added:** Add `Router::layer_on_matching_route` for applying middleware that
+- **added:** Add `Router::route_layer` for applying middleware that
   will only run on requests that match a route ([#474])
 
 [#474]: https://github.com/tokio-rs/axum/pull/474